itergate the api of get and update for dr and pateint and make chnage in the extisting file
