+++ healthhub/src/constants/tailwindConstants.ts
@@ -1,0 +1,4 @@
+ export const ColorClasses = {
+  black: 'text-[#121417]',
+  grayBg: 'bg-[#F0F2F5]',
+};
